tree_recognition/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.5
fonttools==4.54.1
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.1
packaging==24.1
pillow==10.4.0
protobuf==5.28.2
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torchsummary==1.5.1
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.4
